% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_explanation}
\alias{fit_explanation}
\title{Fit white box model to the simulated data.}
\usage{
fit_explanation(live_object, kernel = gaussian_kernel,
  response_family = "gaussian")
}
\arguments{
\item{live_object}{List return by add_predictions function.}

\item{kernel}{function which will be used to calculate distance between simulated
observations and explained instance.}

\item{response_family}{family argument to glmnet (and then glm) function.
Default value is "gaussian"}
}
\value{
List consting of
\item{data}{Dataset used to fit explanation model (may have less column than the original)}
\item{model}{Fitted explanation model}
\item{explained_instance}{Instance that is being explained}
}
\description{
Fit white box model to the simulated data.
}
\examples{
\dontrun{
fitted_explanation <- fit_explanation2(local_exploration1, "regr.lm", selection = TRUE)
}

}
