% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretable_inputs.R
\name{\%:::\%}
\alias{\%:::\%}
\title{Import hidden object without CRAN notes.}
\usage{
pkg \%:::\% name
}
\arguments{
\item{pkg}{Package name}

\item{name}{Function name}
}
\description{
Import hidden object without CRAN notes.
}
\author{
Krystian Igras
}
