% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_surrogate.R
\name{individual_surrogate_model}
\alias{individual_surrogate_model}
\title{LIME-like explanations based on Ceteris Paribus curves}
\usage{
individual_surrogate_model(x, new_observation, size, seed = NULL,
  sampling = "uniform")
}
\arguments{
\item{x}{an explainer created with function DALEX2::explain().}

\item{new_observation}{an observation to be explained. Columns in should correspond to columns in the data argument to x.}

\item{size}{number of similar observation to be sampled.}

\item{seed}{If not NULL, seed will be set to this value for reproducibility.}

\item{sampling}{Parameter that controls sampling while creating new observations.}
}
\value{
data.frame of class local_surrogate_explainer
}
\description{
LIME-like explanations based on Ceteris Paribus curves
}
