% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{local_approximation}
\alias{local_approximation}
\title{Local surrogate model}
\usage{
local_approximation(explainer, observation, size, seed = NULL,
  response_family = "gaussian", n_points = nrow(explainer$data),
  smoothness = 0.3, ...)
}
\arguments{
\item{explainer}{DALEX explainer object.}

\item{observation}{Instance to be explained.}

\item{size}{Number of instance to simulate.}

\item{seed}{Argument to set.seed, if not specified, results will not be reproducible.}

\item{response_family}{Family argument to glmnet function.}

\item{smoothness}{Span argument to loess function.}

\item{n_rows}{Number of points, at which ceteris paribus curves will be calculated.}
}
\value{
glmnet model object
}
\description{
Local surrogate model
}
