% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_print.R
\name{plot_interpretable_feature}
\alias{plot_interpretable_feature}
\title{Plot Ceteris Paribus Profile and discretization}
\usage{
plot_interpretable_feature(x, variable)
}
\arguments{
\item{x}{local_surrogate_explainer object}

\item{variable}{chr, name of the variable to be plotted}
}
\value{
ggplot2 object
}
\description{
Plot Ceteris Paribus Profile and discretization
}
