% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_surrogate.R
\name{predict_model}
\alias{predict_model}
\alias{predict_model.dalex_explainer}
\title{Predict model for lime}
\usage{
\method{predict_model}{dalex_explainer}(x, newdata, ...)
}
\description{
Predict model for lime
}
