% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_surrogate.R
\name{print.local_surrogate_explainer}
\alias{print.local_surrogate_explainer}
\title{Generic print function for local surrogate explainers}
\usage{
\method{print}{local_surrogate_explainer}(x, ...)
}
\arguments{
\item{x}{object of class local_surrogate_explainer}

\item{...}{currently ignored}
}
\description{
Generic print function for local surrogate explainers
}
\examples{
\dontrun{
# Example based on apartments data from DALEX package.
mrf <- randomForest(m2.price ~., data = apartments, ntree = 50)
explainer <- DALEX::explain(model = mrf,
                            data = apartmentsTest[, -1])
model_lok <- individual_surrogate_model(explainer, new_observation,
                                        size = 500, seed = 17)
model_lok
}

}
