% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_surrogate.R
\name{individual_surrogate_model}
\alias{individual_surrogate_model}
\title{LIME-like explanations based on Ceteris Paribus curves}
\usage{
individual_surrogate_model(x, new_observation, size, seed = NULL,
  sampling = "uniform", grid_points = 101)
}
\arguments{
\item{x}{an explainer created with the function DALEX2::explain().}

\item{new_observation}{an observation to be explained. Columns in should correspond to columns in the data argument to x.}

\item{size}{number of similar observation to be sampled.}

\item{seed}{If not NULL, seed will be set to this value for reproducibility.}

\item{sampling}{Parameter that controls sampling while creating new observations.}

\item{grid_points}{Number of points to use while calculating Ceteris Paribus profiles.}
}
\value{
data.frame of class local_surrogate_explainer
}
\description{
LIME-like explanations based on Ceteris Paribus curves
}
\examples{
\dontrun{
#' @examples
\dontrun{
# Example based on apartments data from DALEX package.
mrf <- randomForest(m2.price ~., data = apartments, ntree = 50)
explainer <- DALEX::explain(model = mrf,
                            data = apartmentsTest[, -1])
model_lok <- individual_surrogate_model(explainer, new_observation,
                                        size = 500, seed = 17)
}
}

}
