% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{individual_surrogate_model}
\alias{individual_surrogate_model}
\title{Local surrogate model}
\usage{
individual_surrogate_model(x, new_observation, size, levels = "two",
  seed = NULL, kernel = gaussian_kernel, grid_points = nrow(x$data),
  smoothness = NULL, penalty = NULL)
}
\arguments{
\item{x}{an explainer created with function `DALEX2::explain()`}

\item{new_observation}{an observation/observations to be explained.
Columns in should correspond to columns in the data element of explainer.}

\item{size}{Number of instance to simulate.}

\item{levels}{Should two levels of each variable be created ("baseline", "lower"
or "higher") or three levels ("baseline", "lower" and "higher").}

\item{seed}{Argument to set.seed, if not specified, results will not be reproducible.}

\item{kernel}{kernel function which will be used to weight distances between observations.}

\item{grid_points}{Number of points, at which ceteris paribus curves will be calculated.}

\item{smoothness}{Span argument to loess function. If NULL, smoothing won't be applied
to results of ceteris_paribus function.}

\item{Sequence}{of lambdas (penalties) to glmnet.
If NULL, penalty will be chosen via cross validation.}
}
\value{
glmnet model object
}
\description{
Local surrogate model
}
