% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{individual_surrogate_model}
\alias{individual_surrogate_model}
\title{Local surrogate model}
\usage{
individual_surrogate_model(x, new_observation, size, seed = NULL,
  response_family = "gaussian", n_points = nrow(x$data),
  smoothness = 0.3, ...)
}
\arguments{
\item{x}{an explainer created with function `DALEX2::explain()`}

\item{new_observation}{an observation/observations to be explained.
Columns in should correspond to columns in the data element of explainer.}

\item{size}{Number of instance to simulate.}

\item{seed}{Argument to set.seed, if not specified, results will not be reproducible.}

\item{response_family}{Family argument to glmnet function.}

\item{n_points}{Number of points, at which ceteris paribus curves will be calculated.}

\item{smoothness}{Span argument to loess function.}
}
\value{
glmnet model object
}
\description{
Local surrogate model
}
